+++ index.js
@@ -266,7 +266,7 @@ var _toString2 = _interopRequireDefault(_toString);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var version = '8.2.0';
+var version = '9.0.0';
 
 var validator = {
   version: version,